Updates paths to point to temp and app data.
Updates all cmds to explicitly pass the pin.pin file and moves all file system interactions to the app data folder or the temp. Comments out checkPin as it had "pin.pin" hardcoded, but appears to be unused.

At this point ALL IO should take place in the system definded temp folder if it's updating the pin file, or in the app data folder if it's reading the pins.